Move token migration script to plugin class.
